init
init

+++ index.js
@@ -76,11 +76,16 @@ client.on("message", async message => {
         if(!(message.channel.id == "703185069354778725")) {
 
             if(!(message.content.startsWith(`${prefix}`))) return;
-            var msg = "**Error:** Please use the <#703168301634945097> channel for bot commands!";
-            message.channel.send(`**Error:** Please use the <#703168301634945097> channel for bot commands!`).then(msg => msg.delete({timeout: 10000}));
-            message.delete();
             
-            return;
+            if(!(message.member.hasPermission("ADMINISTRATOR"))) {
+                
+                var msg = "**Error:** Please use the <#703168301634945097> channel for bot commands!";
+                message.channel.send(`**Error:** Please use the <#703168301634945097> channel for bot commands!`).then(msg => msg.delete({timeout: 10000}));
+                message.delete();
+                
+                return;
+            }
+
         }
     }
